Remove ITerminalInstance.copyLastCommandOutputNiche impl should live in command instead
